2-Aside/requirements.txt
# 2-Aside Platform - Shared Python Dependencies
# Install with: pip install -r requirements.txt

# ========================================
# WEB FRAMEWORK
# ========================================
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# ========================================
# DATABASE
# ========================================
sqlalchemy==2.0.27
alembic==1.13.1

# SQL Server drivers
pyodbc==5.1.0
pymssql==2.2.11

# PostgreSQL (alternative)
# psycopg2-binary==2.9.9

# ========================================
# VALIDATION & SERIALIZATION
# ========================================
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# ========================================
# AUTHENTICATION & SECURITY
# ========================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
bcrypt==4.1.2

# ========================================
# CACHING & QUEUE
# ========================================
redis==5.0.1
celery==5.3.6
celery-beat==2.5.0

# ========================================
# AZURE SDK
# ========================================
azure-storage-blob==12.19.0
azure-identity==1.15.0

# ========================================
# HTTP CLIENT
# ========================================
httpx==0.26.0
aiohttp==3.9.3

# ========================================
# TESTING
# ========================================
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==22.6.0
httpx==0.26.0  # For TestClient

# ========================================
# CODE QUALITY
# ========================================
black==24.1.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0

# ========================================
# LOGGING & MONITORING
# ========================================
loguru==0.7.2
python-json-logger==2.0.7

# Optional: Sentry for error tracking
# sentry-sdk==1.40.0

# Optional: Azure Application Insights
# opencensus-ext-azure==1.1.13

# ========================================
# UTILITIES
# ========================================
python-dotenv==1.0.1
pyyaml==6.0.1
pytz==2024.1
python-dateutil==2.8.2

# ========================================
# DEVELOPMENT TOOLS
# ========================================
watchfiles==0.21.0  # For hot reload
ipython==8.21.0
ipdb==0.13.13
